Extract home and lobby into separate components
